requirements.txt
# requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-decouple
psycopg2-binary
gunicorn
whitenoise